fix: item's background color is now consistent
Changing the color only occured when the card was created, but it didn't change when the inventory/equipment's state changed. By binding the color to the item's rarity directly, we avoid that issue (thankfully the CSS variable names match the rarity names).
I also added a little bit of a bottom margin to the Item History button, so it looks nicer and set min-width of the card to 13rem, which looks fine on 1080p screens - it probably isn't a perfect solution.